feat: Update EDA implementation and improve project structure
- Update TFM_EDA.py with enhanced exploratory data analysis
- Add comprehensive logging functionality (log_EDA.txt)
- Add project dependencies specification (requirements.txt)
- Update engine configuration and utilities:
  * Modify engine_TFM/config_eda.yml with new parameters
  * Enhance engine_TFM/utils.py with additional functions
- Improve .gitignore to exclude CSV data files and __pycache__
- Remove compiled Python bytecode files from tracking
- Ensure large datasets (CSV files) are properly ignored

This commit establishes a clean project structure with proper
dependency management and excludes unnecessary files from version control.

+++ engine_TFM/utils.py
@@ -415,7 +415,7 @@ def term_to_int(series: pd.Series, verbose: bool = False) -> pd.Series:
     if series.dtype == int or series.dtype == float:
         return series.astype(float).round().astype(int)
     converted = series.astype(str).str.replace(' months', '', regex=False)
-    converted = converted.str.extract(r"(\d+)")[0].astype(float).round().astype('Int64')
+    converted = converted.str.extract(r"(\d+)")[0].astype(float).round().astype(float)
     if verbose:
         print("[FEATURE] Converted 'term' to integers (nullable).")
     return converted
